+++ Modules/Intent.Modules.EntityFrameworkCore.Repositories/Templates/Repository/RepositoryTemplateRegistration.cs
@@ -41,7 +41,8 @@ namespace Intent.Modules.EntityFrameworkCore.Repositories.Templates.Repository
         public override IEnumerable<ClassModel> GetModels(IApplication application)
         {
             return _metadataManager.Domain(application).GetClassModels()
-                .Where(x => (x.IsAggregateRoot() && (!x.IsAbstract || x.HasTable())) || x.HasRepository())
+                .Where(x => (x.InternalElement?.Package?.SpecializationTypeId == DomainPackageModel.SpecializationTypeId &&
+                             x.IsAggregateRoot() && (!x.IsAbstract || x.HasTable())) || x.HasRepository())
                 .ToArray();
         }
     }