+++ Modules/Intent.Modules.EntityFrameworkCore/Templates/DbContextInterface/DbContextInterfaceTemplateRegistration.cs
@@ -44,7 +44,9 @@ namespace Intent.Modules.EntityFrameworkCore.Templates.DbContextInterface
         [IntentManaged(Mode.Merge, Body = Mode.Ignore, Signature = Mode.Fully)]
         public override IList<ClassModel> GetModels(IApplication application)
         {
-            return _metadataManager.Domain(application).GetClassModels().ToList();
+            return _metadataManager.Domain(application).GetClassModels()
+                .Where(p => p.InternalElement?.Package?.SpecializationTypeId == DomainPackageModel.SpecializationTypeId)
+                .ToList();
         }
     }
 }