+++ Modules/Intent.Modules.EntityFrameworkCore/Templates/EntityTypeConfiguration/EntityTypeConfigurationTemplateRegistration.cs
@@ -40,7 +40,8 @@ namespace Intent.Modules.EntityFrameworkCore.Templates.EntityTypeConfiguration
         public override IEnumerable<ClassModel> GetModels(IApplication application)
         {
             return _metadataManager.Domain(application).GetClassModels()
-                .Where(x => !x.IsOwned(application) && (
+                .Where(x => x.InternalElement?.Package?.SpecializationTypeId == DomainPackageModel.SpecializationTypeId && 
+                    !x.IsOwned(application) && (
                     !x.IsAbstract || // is concrete class
                     x.HasTable() || // has Table stereotype
                     x.AssociatedFromClasses().Any(x => x.IsNullable || x.IsCollection))) // is referenced by other entities